+++ tests/test_clients/test_methods/test_errors/test_messaging.py
@@ -12,12 +12,11 @@ pytest_plugins = ("tests.test_clients.fixtures",)
 
 
 async def test_raising_messaging_error(client, requests_client):
-    method = Info(group_chat_id=uuid.uuid4())
+    method = Info(host="example.com", group_chat_id=uuid.uuid4())
 
     errors_to_raise = {Info: (HTTPStatus.BAD_REQUEST, {})}
 
     with client.error_client(errors=errors_to_raise):
-        method.host = "example.com"
         request = requests_client.build_request(method)
         response = await callable_to_coroutine(requests_client.execute, request)