+++ tests/test_clients/test_methods/test_v3/test_chats/test_info.py
@@ -9,9 +9,8 @@ pytestmark = pytest.mark.asyncio
 
 
 async def test_retrieving_info(client, requests_client):
-    method = Info(group_chat_id=uuid.uuid4())
+    method = Info(host="example.com", group_chat_id=uuid.uuid4())
 
-    method.host = "example.com"
     request = requests_client.build_request(method)
     response = await callable_to_coroutine(requests_client.execute, request)
     info = await callable_to_coroutine(