+++ tests/test_clients/test_methods/test_v3/test_command/test_command_result.py
@@ -12,12 +12,12 @@ pytest_plugins = ("tests.test_clients.fixtures",)
 
 async def test_sending_command_result(client, requests_client):
     method = CommandResult(
+        host="example.com",
         sync_id=uuid.uuid4(),
         bot_id=uuid.uuid4(),
         result=ResultPayload(body="test"),
     )
 
-    method.host = "example.com"
     request = requests_client.build_request(method)
     assert await callable_to_coroutine(requests_client.execute, request)