+++ tests/test_clients/test_methods/test_v3/test_users/test_by_login.py
@@ -8,9 +8,8 @@ pytest_plugins = ("tests.test_clients.fixtures",)
 
 
 async def test_search_by_huid(client, requests_client):
-    method = ByLogin(ad_login="test", ad_domain="example.com")
+    method = ByLogin(host="example.com", ad_login="test", ad_domain="example.com")
 
-    method.host = "example.com"
     request = requests_client.build_request(method)
     response = await callable_to_coroutine(requests_client.execute, request)
     user = await callable_to_coroutine(