+++ tests/test_exceptions/test_route_deprecated_error.py
@@ -12,11 +12,10 @@ pytest_plugins = ("tests.test_clients.fixtures",)
 
 
 async def test_raising_route_deprecated_error(client, requests_client):
-    method = Info(group_chat_id=uuid.uuid4())
+    method = Info(host="example.com", group_chat_id=uuid.uuid4())
     errors_to_raise = {Info: (httpx.codes.GONE, {})}
 
     with client.error_client(errors=errors_to_raise):
-        method.host = "example.com"
         request = requests_client.build_request(method)
         response = await callable_to_coroutine(requests_client.execute, request)